Resolve builtin multiarch Docker variables
Docker exposes various variables related to multi-platform building
in its builds.[1]

dockerfile-json currently does not account for these variables and does
not resolve them properly.

For example:
`FROM BASE-${TARGETARCH}` resolves to `BASE-` but should resolve to the
actual architecture that is used like `BASE-amd64`

Detect the user's arch and OS and expose it via the variables. Let users
overwrite it via build args if they need to.

[1]: https://docs.docker.com/build/building/multi-platform/

Signed-off-by: Tomáš Nevrlka <tnevrlka@redhat.com>