Fix metaArgsEnvExpander
The method returns a function meant to be used as the SingleWordExpander
function. But the behavior of the function does not match what a
SingleWordExpander should do.

What it *should* do:

    "some string with ${var1} and ${var2}"
    -> some string with <var1 value or empty> and <var2 value or empty>

What it actually does:

    "var1" -> <var1 value or error>

Because of this, variable expansion didn't really work at all.

Fix the returned function to behave as intended. However, note that this
DOES NOT fix variable expansion, it just illustrates how broken it was
in the first place.

In the test data, notice that variables that should be getting expanded
to actual values are now getting expanded to empty strings, while some
that should be empty strings are wrongly expanding to values! That's
because we're using the values of meta args (global-only variables) as
local variables in build stages, where meta args should NOT be visible.
And conversely, we have no handling for local variables (ARG/ENV)
defined in the build stages.

The problem above will be fixed in a subsequent commit.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>