Inject the --env variables
This is done to make the behaviour consistent with podman or buildah.
The --env variable is now not used only for variable expansion, but the
env can be found in the resulting parsed json.

As with podman and buildah, the ENV variable already defined in
Dockerfile takes precedence, but the ENV variable define via --cli is
always injected at the start of each stage.

STONEBLD-3890

Signed-off-by: mkosiarc <mkosiarc@redhat.com>